requirements.txt
blinker==1.7.0
Bootstrap-Flask==2.3.3
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
dominate==2.9.1
email-validator==2.1.0.post1
Flask==3.0.0
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2
